Desktop/Code Projects/Patone/python/requirements.txt
# Roadside Assistance Admin Platform - Python Dependencies
# Install with: pip install -r requirements.txt

# Database connectivity
mysql-connector-python==8.0.33

# Data analysis and manipulation
pandas==2.0.3
numpy==1.24.3

# Report generation
reportlab==4.0.4
matplotlib==3.7.2
seaborn==0.12.2

# HTTP requests and API
requests==2.31.0

# Email functionality
smtplib3==0.1.0

# SMS functionality
twilio==8.5.0

# Geolocation services
geopy==2.3.0

# Excel file processing
openpyxl==3.1.2

# PDF processing
PyPDF2==3.0.1

# Image processing
Pillow==10.0.0

# Data validation
pydantic==2.1.1

# Configuration management
python-dotenv==1.0.0

# Logging and monitoring
loguru==0.7.0

# Task scheduling
schedule==1.2.0

# Web scraping (for data import)
beautifulsoup4==4.12.2

# Testing
pytest==7.4.0

# Development tools
black==23.7.0
flake8==6.0.0
